api/requirements/api_requirements.txt
pydantic
pydantic[email]
orjson
asyncpg
passlib[bcrypt]
pyjwt
python-multipart
fastapi
uvicorn[standard]
jinja2
emails